Remove lib folder and move all commands to the main file
