--- lib/i18n.rb
@@ -1,42 +1,0 @@
-def setup_rails_i18n
-  inject_into_class 'app/controllers/application_controller.rb', 'ApplicationController' do
-    "  include Localization\n"
-  end
-
-  gsub_file 'app/views/layouts/application.html.erb', /<html>/ do
-    "<html lang='<%= I18n.locale %>'>"
-  end
-end
-
-def setup_i18n_js
-  environment(nil, env: 'development') do
-    <<~EOT
-      # Automatically generate the `translation.js` files
-      config.middleware.use I18n::JS::Middleware
-
-    EOT
-  end
-
-  insert_into_file 'app/javascript/packs/application.js', after: "require\(\"channels\"\)\n" do
-    <<~EOT
-
-    import 'translations/translations';
-    EOT
-  end
-
-  insert_into_file 'package.json', after: %r{"@rails/ujs": .+\n} do
-    <<~EOT
-    "i18n-js": "^3.0.11",
-    EOT
-  end
-
-  append_to_file '.gitignore' do
-    <<~EOT
-
-      # Ignore i18n.js generated files
-      # If deploy to heroku with git, please remove this as it prevents the files to be committed
-      /app/javascript/translations/translations.js
-    EOT
-  end
-end
-
