+++ template.rb
@@ -5,13 +5,6 @@ APP_NAME = app_name
 RUBY_VERSION = '2.6.5'
 
 def apply_template!
-  apply 'lib/config.rb'
-  apply 'lib/rspec.rb'
-  apply 'lib/test_env.rb'
-  apply 'lib/linter.rb'
-  apply 'lib/bullet.rb'
-  apply 'lib/i18n.rb'
-
   # TODO: Recheck if we need this
   delete_git
   delete_test_folder
@@ -20,7 +13,11 @@ def apply_template!
   directory 'rails_docker/bin', 'bin', mode: :preserve
   directory 'shared', './', force: true, recursive: false
 
-  setup_config
+  # Setup config
+  remove_file 'config/database.yml'
+  directory 'shared/config', 'config',
+  add_sidekiq_config
+  add_mailer_config
 
   # Setup Javascripts
   directory 'shared/app/javascript', 'app/javascript'
@@ -42,8 +39,7 @@ def apply_template!
   remove_turbolinks
 
   setup_rack_deflater
-  setup_rails_i18n
-  setup_i18n_js
+  setup_i18n
   setup_test_env
   setup_bullet
   setup_linters
@@ -56,6 +52,37 @@ def apply_template!
       config.assets.compile = false
 
     EOT
+  end
+
+  after_bundle do
+    run 'spring stop'
+
+    # Devise configuration
+    generate 'devise:install'
+
+    # Add i18n-js plugin to webpack
+    insert_into_file 'config/webpack/environment.js', after: "const { environment } = require('@rails/webpacker')\n" do
+      <<~EOT
+      const webpack = require('webpack');
+      
+      const plugins = [
+        new webpack.ProvidePlugin({
+          // Translations
+          I18n: 'i18n-js',
+        })
+      ]
+  
+      environment.config.set('plugins', plugins);
+      EOT
+    end
+
+    # Setup rspec
+    generate "rspec:install"
+    copy_file 'shared/rspec/.rspec', '.rspec', force: true
+
+    FileUtils.cp_r "#{current_directory}/shared/rspec/.", 'spec/'
+    # remove the .rspec file from rspec folder
+    run 'rm -rf spec/.rspec'
   end
 end
 
@@ -96,11 +123,142 @@ def remove_turbolinks
   gsub_file('package.json', %r{"turbolinks": .+\n}, '')
 end
 
+def add_sidekiq_config
+  insert_into_file 'config/application.rb', after: %r{config.generators.system_tests.+\n} do
+    <<-EOT
+    
+    # Set the queuing backend to `Sidekiq`
+    # 
+    # Be sure to have the adapter's gem in your Gemfile
+    # and follow the adapter's specific installation
+    # and deployment instructions.
+    config.active_job.queue_adapter = :sidekiq
+
+    # Prefix the queue name of all jobs with Rails ENV
+    config.active_job.queue_name_prefix = Rails.env
+    EOT
+  end
+end
+
+def add_mailer_config
+  insert_into_file 'config/environments/development.rb', after: %r{config.action_mailer.perform_caching.+\n} do
+    <<-EOT
+
+  config.action_mailer.default_url_options = { 
+    host: ENV.fetch('MAILER_DEFAULT_HOST'), 
+    port: ENV.fetch('MAILER_DEFAULT_PORT')
+  }
+    EOT
+  end
+end
+
+def setup_i18n
+  inject_into_class 'app/controllers/application_controller.rb', 'ApplicationController' do
+    "  include Localization\n"
+  end
+
+  gsub_file 'app/views/layouts/application.html.erb', /<html>/ do
+    "<html lang='<%= I18n.locale %>'>"
+  end
+
+  environment(nil, env: 'development') do
+    <<~EOT
+      # Automatically generate the `translation.js` files
+      config.middleware.use I18n::JS::Middleware
+
+    EOT
+  end
+
+  insert_into_file 'app/javascript/packs/application.js', after: "require\(\"channels\"\)\n" do
+    <<~EOT
+
+    import 'translations/translations';
+    EOT
+  end
+
+  insert_into_file 'package.json', after: %r{"@rails/ujs": .+\n} do
+    <<~EOT
+    "i18n-js": "^3.0.11",
+    EOT
+  end
+
+  append_to_file '.gitignore' do
+    <<~EOT
+
+      # Ignore i18n.js generated files
+      # If deploy to heroku with git, please remove this as it prevents the files to be committed
+      /app/javascript/translations/translations.js
+    EOT
+  end
+end
+
 def setup_rack_deflater
   environment do
     <<~EOT
       # Compress the responses to reduce the size of html/json controller responses.
       config.middleware.use Rack::Deflater
+
+    EOT
+  end
+end
+
+def setup_bullet
+  environment(nil, env: 'development') do
+    <<~EOT
+      # Configure Bullet gem to detect N+1 queries
+      config.after_initialize do
+        Bullet.enable        = true
+        Bullet.bullet_logger = true
+        Bullet.console       = true
+        Bullet.rails_logger  = true
+        Bullet.add_footer    = true
+      end
+
+    EOT
+  end
+
+  environment(nil, env: 'test') do
+    <<~EOT
+      # Configure Bullet gem to detect N+1 queries
+      config.after_initialize do
+        Bullet.enable                      = true
+        Bullet.bullet_logger               = true
+        Bullet.raise                       = true
+        Bullet.unused_eager_loading_enable = false
+      end
+
+    EOT
+  end
+end
+
+def setup_linters
+  directory 'shared/linters', './'
+
+  insert_into_file 'package.json', before: %r{"version": .+\n} do
+    <<~EOT
+      "devDependencies": { 
+        "@nimbl3/eslint-config-nimbl3": "2.1.1"
+      },
+    EOT
+  end
+end
+
+def setup_test_env
+  disable_animations
+end
+
+def disable_animations
+  insert_into_file 'config/environments/test.rb', before: %r{Rails.application.configure do} do
+    <<~EOT
+      require_relative '../../spec/support/disable_animation'
+
+    EOT
+  end
+
+  environment(nil, env: 'test') do
+    <<~EOT
+      # Disable all animation during tests
+      config.middleware.use Rack::NoAnimations
 
     EOT
   end
@@ -121,40 +279,6 @@ def setup_gitignore
       .pronto.yml
     EOT
   end
-end
-
-after_bundle do
-  run 'spring stop'
-
-  # Devise configuration
-  generate 'devise:install'
-  insert_into_file 'config/environments/development.rb', after: %r{config.action_mailer.perform_caching.+\n} do
-    <<-EOT
-
-  config.action_mailer.default_url_options = { 
-    host: ENV.fetch('MAILER_DEFAULT_HOST'), 
-    port: ENV.fetch('MAILER_DEFAULT_PORT')
-  }
-    EOT
-  end
-
-  # Add i18n-js plugin to webpack
-  insert_into_file 'config/webpack/environment.js', after: "const { environment } = require('@rails/webpacker')\n" do
-    <<~EOT
-      const webpack = require('webpack');
-      
-      const plugins = [
-        new webpack.ProvidePlugin({
-          // Translations
-          I18n: 'i18n-js',
-        })
-      ]
-  
-      environment.config.set('plugins', plugins);
-    EOT
-  end
-
-  setup_rspec
 end
 
 apply_template!
